+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.2.4 (under development)
 
 - Enhanced `Nav` to allow for `active` callables
+- Enhanced `Navbar` item sorting by introducing an optional `order` array key
 - Enhanced `SignupForm` to better handle the token validation
 
 ## 2.2.3 (Mar 4, 2024)